+++ src/controllers/warehouse_Product/deleteWHProducts.ts
