+++ src/services/warehouse_product/index.ts
@@ -1,10 +1,10 @@
-import createWarehouseProduct from './serviceCreate'
-import getAllWarehouseProducts from './serviceGetAll'
-import getWarehouseProductById from './serviceGetById'
-import getWarehouseProductsByWarehouse from './serviceGetByWarehouse'
-import getWarehouseProductsByProduct from './serviceGetByProduct'
-import updateWarehouseProduct from './serviceUpdate'
-import deleteWarehouseProduct from './serviceDelete'
+import createWarehouseProduct from './serviceCreateWHProducts'
+import getAllWarehouseProducts from './serviceGetAllWHProducts'
+import getWarehouseProductById from './serviceGetByIdWHProducts'
+import getWarehouseProductsByWarehouse from './serviceGetByWarehouse_WHProducts'
+import getWarehouseProductsByProduct from './serviceGetByProductWHProducts'
+import updateWarehouseProduct from './serviceUpdateWHProducts'
+import deleteWarehouseProduct from './serviceDeleteWHProducts'
 
 const warehouseProductService = {
   createWarehouseProduct,
