+++ src/services/warehouse_product/serviceDeleteWHProducts.ts
